Move Examples to root (#2176)
1. Moves the examples to root
2. Accordingly calls out the examples as WIP in the README
3. Uses absolute instead of relative import now

This is a cascading PR. Please review and merge
https://github.com/explodinggradients/ragas/pull/2173 before reviewing
this.
