+++ examples/README.md
@@ -1,0 +1,11 @@
+# Ragas Examples for Contributors
+
+This work in progress directory will contain comprehensive examples demonstrating how to use Ragas for evaluating different types of AI applications including RAG systems, agents, prompts, and workflows. These examples might be unstable and are subject to change.
+
+## Installation
+
+Install the ragas package with examples dependencies:
+
+```bash
+uv pip install --system -e ".[examples]"
+```
