+++ examples/prompt_evals/evals.py
@@ -2,7 +2,7 @@ from ragas.experimental import Dataset, experiment
 from ragas.experimental.metrics.result import MetricResult
 from ragas.experimental.metrics.discrete import discrete_metric
 
-from .prompt import run_prompt
+from prompt import run_prompt
 
 
 @discrete_metric(name="accuracy", allowed_values=["pass", "fail"])
