+++ examples/rag_eval/evals.py
@@ -3,7 +3,7 @@ from ragas.experimental.metrics import DiscreteMetric
 from openai import OpenAI
 from ragas.experimental.llms import llm_factory
 import os
-from .rag import default_rag_client
+from rag import default_rag_client
 
 openai_client = OpenAI(api_key=os.environ.get("OPENAI_API_KEY"))
 rag_client = default_rag_client(llm_client=openai_client)
