--- ragas/examples/experimental/README.md
@@ -1,12 +1,0 @@
-# Ragas Examples
-
-This package contains comprehensive examples demonstrating how to use Ragas for evaluating different types of AI applications including RAG systems, agents, prompts, and workflows.
-
-## Installation
-
-Install the ragas_experimental package with examples dependencies:
-
-```bash
-pip install -e ".[examples]"
-```
-
